put the stringHash function in a separate file
